spinalcordtoolbox/cropping: Moved ImageCropper into python module
